Split test into separate binaries
